fix warnings
Former-commit-id: e6135c4d5463e42251867e57001414dbf31ef12f [formerly de4f8f624413879b8d95ab1e9ea9d81d5e679de1] [formerly 6d4fc41a5093219a946a68d43f9b1dc5035d4dc9 [formerly dd23f61663f621bffb55a9ffbb7f605d4d62d3e3]]
Former-commit-id: e0114562644db8967553dfb2766c840247bf93f5 [formerly ca13547a8cc47ef42d770e79426fb5e2c828c8c1]
Former-commit-id: 0c501c3418bdb5d31b211d8ec0601d9376002294

+++ Frame/Server/WorldServer/WorldNetServerPlugin/AFCWorldNetServerModule.cpp
@@ -400,7 +400,7 @@ bool AFCWorldNetServerModule::SendMsgToGame(const AFIDataList& argObjectVar, con
         return false;
     }
 
-    for(int i = 0; i < argObjectVar.GetCount(); i++)
+    for(size_t i = 0; i < argObjectVar.GetCount(); i++)
     {
         const AFGUID& identOther = argObjectVar.Object(i);
         const int64_t  nGameID = argGameID.Int(i);
@@ -430,7 +430,7 @@ int AFCWorldNetServerModule::OnObjectListEnter(const AFIDataList& self, const AF
     }
 
     AFMsg::AckPlayerEntryList xPlayerEntryInfoList;
-    for(int i = 0; i < argVar.GetCount(); i++)
+    for(size_t i = 0; i < argVar.GetCount(); i++)
     {
         AFGUID identOld = argVar.Object(i);
         //排除空对象
@@ -455,7 +455,7 @@ int AFCWorldNetServerModule::OnObjectListEnter(const AFIDataList& self, const AF
         return 0;
     }
 
-    for(int i = 0; i < self.GetCount(); i++)
+    for(size_t i = 0; i < self.GetCount(); i++)
     {
         AFGUID ident = self.Object(i);
         if(ident.IsNULL())
@@ -479,7 +479,7 @@ int AFCWorldNetServerModule::OnObjectListLeave(const AFIDataList& self, const AF
     }
 
     AFMsg::AckPlayerLeaveList xPlayerLeaveInfoList;
-    for(int i = 0; i < argVar.GetCount(); i++)
+    for(size_t i = 0; i < argVar.GetCount(); i++)
     {
         AFGUID identOld = argVar.Object(i);
         //排除空对象
@@ -492,7 +492,7 @@ int AFCWorldNetServerModule::OnObjectListLeave(const AFIDataList& self, const AF
         *pIdent = AFINetServerModule::GUIDToPB(argVar.Object(i));
     }
 
-    for(int i = 0; i < self.GetCount(); i++)
+    for(size_t i = 0; i < self.GetCount(); i++)
     {
         AFGUID ident = self.Object(i);
         if(ident.IsNULL())
@@ -532,8 +532,8 @@ int AFCWorldNetServerModule::OnDataTableEnter(const AFIDataList& argVar, const A
     AFMsg::ObjectRecordList* pPrivateData = NULL;
 
     ARK_SHARE_PTR<AFIDataTableManager> pTableManager = pEntity->GetTableManager();
-    int nTableCount = pTableManager->GetCount();
-    for(int i = 0; i < nTableCount; ++i)
+    size_t nTableCount = pTableManager->GetCount();
+    for(size_t i = 0; i < nTableCount; ++i)
     {
         AFDataTable* pTable = pTableManager->GetTableByIndex(i);
         if(pTable == nullptr)
@@ -575,7 +575,7 @@ int AFCWorldNetServerModule::OnDataTableEnter(const AFIDataList& argVar, const A
         }
     }
 
-    for(int i = 0; i < argVar.GetCount(); i++)
+    for(size_t i = 0; i < argVar.GetCount(); i++)
     {
         const AFGUID& identOther = argVar.Object(i);
         const int64_t nGameID = argGameID.Int(i);
